Refactor GhCanvasTool into partial classes
Split the 1,977-line GhCanvasTool.cs into 4 files for better
maintainability and code navigation:

- GhCanvasTool.cs: Core dispatch, ToolInfo, and main canvas actions
- GhCanvasTool.Values.cs: Value adjustment actions (get, set, config, preview, enable)
- GhCanvasTool.Groups.cs: Group management actions (create, delete, add, remove, list, etc.)
- GhCanvasTool.Zoomable.cs: Zoomable parameter management (add/remove parameters)

No functional changes - the MCP tool API remains identical.
