+++ src/Cordyceps/Tools/Unified/GhCanvasTool.cs
@@ -16,7 +16,7 @@ namespace Cordyceps.Tools.Unified
     /// Unified canvas tool - component operations, value adjustments, and group management
     /// </summary>
     [McpServerToolType]
-    public class GhCanvasTool
+    public partial class GhCanvasTool
     {
         private readonly GrasshopperContext _context;
 
@@ -1121,857 +1121,8 @@ namespace Cordyceps.Tools.Unified
             return null;
         }
 
-        #region Value Adjustment Actions (from gh_adjust)
-
-        private string ActionGet(string id)
-        {
-            return _context.ExecuteOnUiThread(() =>
-            {
-                if (!ToolHelpers.TryGetUnprotectedComponent(_context, id, out var component, out var error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                object currentValue = null;
-                string valueType = "unknown";
-
-                if (component is GH_NumberSlider slider)
-                {
-                    currentValue = slider.CurrentValue;
-                    valueType = "slider";
-                }
-                else if (component is GH_Panel panel)
-                {
-                    currentValue = panel.UserText;
-                    valueType = "panel";
-                }
-                else if (component is GH_BooleanToggle toggle)
-                {
-                    currentValue = toggle.Value;
-                    valueType = "toggle";
-                }
-                else if (component is GH_ValueList valueList)
-                {
-                    var selected = valueList.SelectedItems.FirstOrDefault();
-                    currentValue = selected?.Name ?? "(none)";
-                    valueType = "value_list";
-                }
-
-                return JsonConvert.SerializeObject(new
-                {
-                    success = true,
-                    id,
-                    name = component.Name,
-                    nickname = component.NickName,
-                    valueType,
-                    value = currentValue
-                });
-            });
-        }
-
-        private string ActionSet(string id, string value)
-        {
-            return _context.ExecuteOnUiThread(() =>
-            {
-                if (!ToolHelpers.TryGetUnprotectedComponentWithDoc(_context, id, out var doc, out var component, out var error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                if (component is GH_Panel panel)
-                {
-                    panel.UserText = value;
-                }
-                else if (component is GH_NumberSlider slider)
-                {
-                    if (!double.TryParse(value, out double numValue))
-                        return ToolHelpers.ErrorResponse($"Invalid number: {value}");
-
-                    if (numValue < (double)slider.Slider.Minimum || numValue > (double)slider.Slider.Maximum)
-                        return ToolHelpers.ErrorResponse($"Value {numValue} out of range [{slider.Slider.Minimum}, {slider.Slider.Maximum}]");
-
-                    slider.SetSliderValue((decimal)numValue);
-                }
-                else if (component is GH_BooleanToggle toggle)
-                {
-                    toggle.Value = value.ToLower() == "true" || value == "1";
-                }
-                else if (component is GH_ValueList valueList)
-                {
-                    var item = valueList.ListItems.FirstOrDefault(i =>
-                        i.Name.Equals(value, StringComparison.OrdinalIgnoreCase) ||
-                        i.Expression.Equals(value, StringComparison.OrdinalIgnoreCase));
-
-                    if (item == null)
-                        return ToolHelpers.ErrorResponse($"Value '{value}' not found in list. Available: {string.Join(", ", valueList.ListItems.Select(i => i.Name))}");
-
-                    valueList.SelectItem(valueList.ListItems.IndexOf(item));
-                }
-                else
-                {
-                    return ToolHelpers.ErrorResponse($"Cannot set value on {component.GetType().Name}");
-                }
-
-                doc.NewSolution(true);
-
-                return JsonConvert.SerializeObject(new
-                {
-                    success = true,
-                    id,
-                    type = component.Name,
-                    value
-                });
-            });
-        }
-
-        private string ActionConfig(string id, double min, double max, string value, int decimals, string items)
-        {
-            return _context.ExecuteOnUiThread(() =>
-            {
-                if (!ToolHelpers.TryGetUnprotectedComponentWithDoc(_context, id, out var doc, out var component, out var error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                if (component is GH_NumberSlider slider)
-                {
-                    if (!double.IsNaN(min)) slider.Slider.Minimum = (decimal)min;
-                    if (!double.IsNaN(max)) slider.Slider.Maximum = (decimal)max;
-                    if (decimals >= 0) slider.Slider.DecimalPlaces = decimals;
-                    if (!string.IsNullOrEmpty(value) && double.TryParse(value, out double val))
-                        slider.SetSliderValue((decimal)val);
-
-                    doc.NewSolution(true);
-
-                    return JsonConvert.SerializeObject(new
-                    {
-                        success = true,
-                        id,
-                        type = "slider",
-                        min = (double)slider.Slider.Minimum,
-                        max = (double)slider.Slider.Maximum,
-                        value = (double)slider.CurrentValue,
-                        decimals = slider.Slider.DecimalPlaces
-                    });
-                }
-                else if (component is GH_ValueList valueList && !string.IsNullOrEmpty(items))
-                {
-                    try
-                    {
-                        var itemList = JsonConvert.DeserializeObject<List<Dictionary<string, string>>>(items);
-                        valueList.ListItems.Clear();
-
-                        foreach (var item in itemList)
-                        {
-                            var itemName = item.ContainsKey("name") ? item["name"] : "Item";
-                            var expr = item.ContainsKey("value") ? item["value"] : "0";
-                            valueList.ListItems.Add(new GH_ValueListItem(itemName, expr));
-                        }
-
-                        if (valueList.ListItems.Count > 0)
-                            valueList.SelectItem(0);
-
-                        doc.NewSolution(true);
-
-                        return JsonConvert.SerializeObject(new
-                        {
-                            success = true,
-                            id,
-                            type = "value_list",
-                            itemCount = valueList.ListItems.Count,
-                            items = valueList.ListItems.Select(i => new { name = i.Name, value = i.Expression }).ToList()
-                        });
-                    }
-                    catch (Exception ex)
-                    {
-                        return ToolHelpers.ErrorResponse($"Failed to parse items: {ex.Message}");
-                    }
-                }
-                else
-                {
-                    return ToolHelpers.ErrorResponse($"Cannot configure {component.GetType().Name}. Use config for sliders and value lists.");
-                }
-            });
-        }
-
-        private string ActionPreview(string id, string ids, bool enabled)
-        {
-            return _context.ExecuteOnUiThread(() =>
-            {
-                if (!ToolHelpers.TryGetActiveDocument(_context, out var doc, out var error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                List<string> idList = BuildIdList(id, ids, out error);
-                if (idList == null) return ToolHelpers.ErrorResponse(error);
-
-                int changed = 0;
-                foreach (var compId in idList)
-                {
-                    if (ToolHelpers.TryGetUnprotectedComponent(_context, compId, out var comp, out _))
-                    {
-                        if (comp is IGH_PreviewObject preview)
-                        {
-                            preview.Hidden = !enabled;
-                            changed++;
-                        }
-                    }
-                }
-
-                Instances.ActiveCanvas?.Invalidate();
-
-                return JsonConvert.SerializeObject(new
-                {
-                    success = true,
-                    previewEnabled = enabled,
-                    changedCount = changed
-                });
-            });
-        }
-
-        private string ActionEnable(string id, string ids, bool enabled)
-        {
-            return _context.ExecuteOnUiThread(() =>
-            {
-                if (!ToolHelpers.TryGetActiveDocument(_context, out var doc, out var error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                List<string> idList = BuildIdList(id, ids, out error);
-                if (idList == null) return ToolHelpers.ErrorResponse(error);
-
-                int changed = 0;
-                foreach (var compId in idList)
-                {
-                    if (ToolHelpers.TryGetUnprotectedComponent(_context, compId, out var comp, out _))
-                    {
-                        if (comp is IGH_ActiveObject active)
-                        {
-                            active.Locked = !enabled;
-                            changed++;
-                        }
-                    }
-                }
-
-                doc.NewSolution(true);
-
-                return JsonConvert.SerializeObject(new
-                {
-                    success = true,
-                    enabled,
-                    changedCount = changed
-                });
-            });
-        }
-
-        private List<string> BuildIdList(string id, string ids, out string error)
-        {
-            error = null;
-            if (!string.IsNullOrEmpty(ids))
-            {
-                try { return JsonConvert.DeserializeObject<List<string>>(ids); }
-                catch { error = "Invalid ids JSON array"; return null; }
-            }
-            if (!string.IsNullOrEmpty(id))
-                return new List<string> { id };
-
-            error = "Either 'id' or 'ids' is required";
-            return null;
-        }
-
-        #endregion
-
-        #region Group Actions (from gh_group)
-
-        private string ActionGroupCreate(string name, string componentIds, string color)
-        {
-            return _context.ExecuteOnUiThread(() =>
-            {
-                if (!ToolHelpers.TryGetActiveDocument(_context, out var doc, out var error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                var group = new GH_Group();
-                group.NickName = name;
-                group.Name = name;
-
-                if (ToolHelpers.TryParseColor(color, out var groupColor))
-                    group.Colour = groupColor;
-
-                doc.AddObject(group, false);
-
-                var addedIds = new List<string>();
-                if (!string.IsNullOrEmpty(componentIds))
-                {
-                    if (ToolHelpers.TryParseGuidArray(componentIds, out var guids, out _))
-                    {
-                        foreach (var guid in guids)
-                        {
-                            var obj = doc.FindObject(guid, true);
-                            if (obj != null)
-                            {
-                                group.AddObject(obj.InstanceGuid);
-                                addedIds.Add(guid.ToString());
-                            }
-                        }
-                        group.ExpireCaches();
-                    }
-                }
-
-                Instances.ActiveCanvas?.Invalidate();
-
-                return JsonConvert.SerializeObject(new
-                {
-                    success = true,
-                    id = group.InstanceGuid.ToString(),
-                    name = group.NickName,
-                    color = ColorTranslator.ToHtml(group.Colour),
-                    memberCount = addedIds.Count,
-                    memberIds = addedIds
-                });
-            });
-        }
-
-        private string ActionGroupDelete(string id)
-        {
-            return _context.ExecuteOnUiThread(() =>
-            {
-                if (!ToolHelpers.TryGetUnprotectedComponentWithDoc(_context, id, out var doc, out var obj, out var error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                if (!(obj is GH_Group group))
-                    return ToolHelpers.ErrorResponse($"Object is not a group: {id}");
-
-                doc.RemoveObject(group, true);
-
-                return JsonConvert.SerializeObject(new
-                {
-                    success = true,
-                    deleted = id
-                });
-            });
-        }
-
-        private string ActionGroupAdd(string groupId, string componentIds, string groupName, string color)
-        {
-            return _context.ExecuteOnUiThread(() =>
-            {
-                if (!ToolHelpers.TryGetActiveDocument(_context, out var doc, out var error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                if (!ToolHelpers.TryParseGuidArray(componentIds, out var guids, out error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                var objects = new List<IGH_DocumentObject>();
-                foreach (var guid in guids)
-                {
-                    var obj = doc.FindObject(guid, true);
-                    if (obj != null)
-                        objects.Add(obj);
-                }
-
-                if (objects.Count == 0)
-                    return ToolHelpers.ErrorResponse("No valid components found");
-
-                GH_Group group = null;
-                if (!string.IsNullOrEmpty(groupId) && Guid.TryParse(groupId, out Guid gGuid))
-                    group = doc.FindObject(gGuid, true) as GH_Group;
-
-                if (group == null)
-                {
-                    group = new GH_Group();
-                    group.NickName = groupName ?? "Group";
-                    group.Name = groupName ?? "Group";
-                    doc.AddObject(group, false);
-                }
-
-                if (ToolHelpers.TryParseColor(color, out var parsedColor))
-                    group.Colour = parsedColor;
-
-                foreach (var obj in objects)
-                    group.AddObject(obj.InstanceGuid);
-
-                group.ExpireCaches();
-                Instances.ActiveCanvas?.Invalidate();
-
-                var bounds = group.Attributes?.Bounds ?? RectangleF.Empty;
-
-                return JsonConvert.SerializeObject(new
-                {
-                    success = true,
-                    groupId = group.InstanceGuid.ToString(),
-                    groupName = group.NickName,
-                    addedCount = objects.Count,
-                    addedIds = objects.Select(o => o.InstanceGuid.ToString()).ToList(),
-                    bounds = new { x = bounds.X, y = bounds.Y, width = bounds.Width, height = bounds.Height }
-                });
-            });
-        }
-
-        private string ActionGroupRemove(string groupId, string componentIds)
-        {
-            return _context.ExecuteOnUiThread(() =>
-            {
-                if (!ToolHelpers.TryGetComponent(_context, groupId, out var obj, out var error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                if (!(obj is GH_Group group))
-                    return ToolHelpers.ErrorResponse($"Object is not a group: {groupId}");
-
-                if (!ToolHelpers.TryParseGuidArray(componentIds, out var guids, out error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                int removedCount = 0;
-                var removedIds = new List<string>();
-
-                foreach (var guid in guids)
-                {
-                    if (group.ObjectIDs != null && group.ObjectIDs.Contains(guid))
-                    {
-                        group.RemoveObject(guid);
-                        removedCount++;
-                        removedIds.Add(guid.ToString());
-                    }
-                }
-
-                group.ExpireCaches();
-                Instances.ActiveCanvas?.Invalidate();
-
-                return JsonConvert.SerializeObject(new
-                {
-                    success = true,
-                    groupId = group.InstanceGuid.ToString(),
-                    removedCount,
-                    removedIds
-                });
-            });
-        }
-
-        private string ActionGroupList()
-        {
-            return _context.ExecuteOnUiThread(() =>
-            {
-                if (!ToolHelpers.TryGetActiveDocument(_context, out var doc, out var error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                var infraIds = ToolHelpers.GetCordycepsInfrastructureIds(doc);
-                var groups = new List<object>();
-
-                foreach (var obj in doc.Objects)
-                {
-                    if (obj is GH_Group group && !infraIds.Contains(group.InstanceGuid))
-                    {
-                        var memberIds = group.ObjectIDs?
-                            .Where(g => !infraIds.Contains(g))
-                            .Select(g => g.ToString())
-                            .ToList() ?? new List<string>();
-
-                        var bounds = group.Attributes?.Bounds ?? RectangleF.Empty;
-
-                        groups.Add(new
-                        {
-                            id = group.InstanceGuid.ToString(),
-                            name = group.NickName,
-                            color = ColorTranslator.ToHtml(group.Colour),
-                            memberCount = memberIds.Count,
-                            memberIds,
-                            bounds = new { x = bounds.X, y = bounds.Y, width = bounds.Width, height = bounds.Height }
-                        });
-                    }
-                }
-
-                return JsonConvert.SerializeObject(new
-                {
-                    success = true,
-                    count = groups.Count,
-                    groups
-                });
-            });
-        }
-
-        private string ActionGroupRename(string id, string newName)
-        {
-            return _context.ExecuteOnUiThread(() =>
-            {
-                if (!ToolHelpers.TryGetComponent(_context, id, out var obj, out var error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                if (!(obj is GH_Group group))
-                    return ToolHelpers.ErrorResponse($"Object is not a group: {id}");
-
-                string oldName = group.NickName;
-                group.NickName = newName;
-                group.Name = newName;
-                Instances.ActiveCanvas?.Invalidate();
-
-                return JsonConvert.SerializeObject(new
-                {
-                    success = true,
-                    id = group.InstanceGuid.ToString(),
-                    oldName,
-                    name = group.NickName
-                });
-            });
-        }
-
-        private string ActionGroupColor(string id, string color)
-        {
-            return _context.ExecuteOnUiThread(() =>
-            {
-                if (!ToolHelpers.TryGetComponent(_context, id, out var obj, out var error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                if (!(obj is GH_Group group))
-                    return ToolHelpers.ErrorResponse($"Object is not a group: {id}");
-
-                if (!ToolHelpers.TryParseColor(color, out var groupColor))
-                    return ToolHelpers.ErrorResponse($"Invalid color: {color}");
-
-                group.Colour = groupColor;
-                Instances.ActiveCanvas?.Invalidate();
-
-                return JsonConvert.SerializeObject(new
-                {
-                    success = true,
-                    id,
-                    color = ColorTranslator.ToHtml(group.Colour)
-                });
-            });
-        }
-
-        private string ActionGroupMove(string groupSpec, double dx, double dy)
-        {
-            return _context.ExecuteOnUiThread(() =>
-            {
-                if (string.IsNullOrEmpty(groupSpec))
-                    return ToolHelpers.ErrorResponse("Group id is required");
-
-                if (!ToolHelpers.TryGetActiveDocument(_context, out var doc, out var error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                var infraIds = ToolHelpers.GetCordycepsInfrastructureIds(doc);
-
-                GH_Group targetGroup = null;
-                if (Guid.TryParse(groupSpec, out Guid groupGuid))
-                {
-                    if (infraIds.Contains(groupGuid))
-                        return ToolHelpers.ErrorResponse($"Protected: required for MCP server. Cannot modify: {groupSpec}");
-                    targetGroup = doc.FindObject(groupGuid, true) as GH_Group;
-                }
-
-                if (targetGroup == null)
-                {
-                    foreach (var obj in doc.Objects)
-                    {
-                        if (obj is GH_Group g && !infraIds.Contains(g.InstanceGuid) &&
-                            (g.NickName.Equals(groupSpec, StringComparison.OrdinalIgnoreCase) ||
-                             g.Name.Equals(groupSpec, StringComparison.OrdinalIgnoreCase)))
-                        {
-                            targetGroup = g;
-                            break;
-                        }
-                    }
-                }
-
-                if (targetGroup == null)
-                    return ToolHelpers.ErrorResponse($"Group not found: {groupSpec}");
-
-                var memberIds = targetGroup.ObjectIDs?.ToList() ?? new List<Guid>();
-                if (memberIds.Count == 0)
-                {
-                    return JsonConvert.SerializeObject(new
-                    {
-                        success = true,
-                        groupId = targetGroup.InstanceGuid.ToString(),
-                        movedCount = 0,
-                        message = "Group has no members"
-                    });
-                }
-
-                int movedCount = 0;
-                foreach (var memberId in memberIds)
-                {
-                    var member = doc.FindObject(memberId, true);
-                    if (member?.Attributes != null)
-                    {
-                        var pivot = member.Attributes.Pivot;
-                        member.Attributes.Pivot = new PointF(pivot.X + (float)dx, pivot.Y + (float)dy);
-                        member.Attributes.ExpireLayout();
-                        movedCount++;
-                    }
-                }
-
-                targetGroup.ExpireCaches();
-                Instances.ActiveCanvas?.Invalidate();
-
-                var bounds = targetGroup.Attributes?.Bounds ?? RectangleF.Empty;
-
-                return JsonConvert.SerializeObject(new
-                {
-                    success = true,
-                    groupId = targetGroup.InstanceGuid.ToString(),
-                    movedCount,
-                    offset = new { dx, dy },
-                    bounds = new { x = bounds.X, y = bounds.Y, width = bounds.Width, height = bounds.Height }
-                });
-            });
-        }
-
-        #endregion
-
-        #region Zoomable Actions
-
-        private string ActionZoomable(string id, string side, string operation, int count, int index)
-        {
-            return _context.ExecuteOnUiThread(() =>
-            {
-                if (!ToolHelpers.TryGetUnprotectedComponentWithDoc(_context, id, out var doc, out var component, out var error))
-                    return ToolHelpers.ErrorResponse(error);
-
-                if (!(component is IGH_Component ghComponent))
-                    return ToolHelpers.ErrorResponse("Component does not support parameters");
-
-                if (!(component is IGH_VariableParameterComponent varParamComp))
-                {
-                    return JsonConvert.SerializeObject(new
-                    {
-                        success = false,
-                        error = $"Component '{ghComponent.NickName}' does not support variable parameters",
-                        componentType = component.GetType().Name,
-                        suggestion = "This action only works with components like Stream Filter, Merge, Entwine, Gate, etc."
-                    });
-                }
-
-                // Default side to input
-                var paramSide = GH_ParameterSide.Input;
-                if (!string.IsNullOrEmpty(side))
-                {
-                    if (side.Equals("output", StringComparison.OrdinalIgnoreCase))
-                        paramSide = GH_ParameterSide.Output;
-                    else if (!side.Equals("input", StringComparison.OrdinalIgnoreCase))
-                        return ToolHelpers.ErrorResponse($"Invalid side: {side}. Use 'input' or 'output'");
-                }
-
-                // Default operation to add
-                var op = operation?.ToLowerInvariant() ?? "add";
-
-                try
-                {
-                    int initialCount = paramSide == GH_ParameterSide.Input
-                        ? ghComponent.Params.Input.Count
-                        : ghComponent.Params.Output.Count;
-
-                    switch (op)
-                    {
-                        case "add":
-                            return AddParameter(ghComponent, varParamComp, paramSide, index, doc);
-
-                        case "remove":
-                            return RemoveParameter(ghComponent, varParamComp, paramSide, index, doc);
-
-                        case "set_count":
-                            if (count < 0)
-                                return ToolHelpers.ErrorResponse("'count' parameter required for set_count operation");
-                            return SetParameterCount(ghComponent, varParamComp, paramSide, count, doc);
-
-                        default:
-                            return ToolHelpers.ErrorResponse($"Invalid operation: {operation}. Use 'add', 'remove', or 'set_count'");
-                    }
-                }
-                catch (Exception ex)
-                {
-                    Core.DebugLog.Error($"gh_canvas zoomable failed: {ex.Message}");
-                    return ToolHelpers.ErrorResponse($"Failed to manage zoomable parameters: {ex.Message}");
-                }
-            });
-        }
-
-        private string AddParameter(IGH_Component ghComponent, IGH_VariableParameterComponent varParamComp,
-            GH_ParameterSide side, int index, GH_Document doc)
-        {
-            var paramList = side == GH_ParameterSide.Input ? ghComponent.Params.Input : ghComponent.Params.Output;
-
-            // If index not specified, add at the end
-            int insertIndex = index >= 0 ? index : paramList.Count;
-
-            if (!varParamComp.CanInsertParameter(side, insertIndex))
-            {
-                return JsonConvert.SerializeObject(new
-                {
-                    success = false,
-                    error = $"Cannot insert parameter at index {insertIndex}",
-                    currentCount = paramList.Count,
-                    suggestion = "Component may have reached maximum parameter count or index is invalid"
-                });
-            }
-
-            var newParam = varParamComp.CreateParameter(side, insertIndex);
-            if (newParam == null)
-            {
-                return ToolHelpers.ErrorResponse("Failed to create parameter");
-            }
-
-            if (side == GH_ParameterSide.Input)
-                ghComponent.Params.RegisterInputParam(newParam);
-            else
-                ghComponent.Params.RegisterOutputParam(newParam);
-
-            varParamComp.VariableParameterMaintenance();
-            ghComponent.ExpireSolution(true);
-            doc.NewSolution(false);
-
-            return JsonConvert.SerializeObject(new
-            {
-                success = true,
-                id = ghComponent.InstanceGuid.ToString(),
-                operation = "add",
-                side = side.ToString().ToLower(),
-                index = insertIndex,
-                parameterName = newParam.Name,
-                newCount = paramList.Count
-            });
-        }
-
-        private string RemoveParameter(IGH_Component ghComponent, IGH_VariableParameterComponent varParamComp,
-            GH_ParameterSide side, int index, GH_Document doc)
-        {
-            var paramList = side == GH_ParameterSide.Input ? ghComponent.Params.Input : ghComponent.Params.Output;
-
-            // If index not specified, remove from the end
-            int removeIndex = index >= 0 ? index : paramList.Count - 1;
-
-            if (removeIndex < 0 || removeIndex >= paramList.Count)
-            {
-                return JsonConvert.SerializeObject(new
-                {
-                    success = false,
-                    error = $"Invalid index {removeIndex}",
-                    currentCount = paramList.Count,
-                    suggestion = $"Index must be between 0 and {paramList.Count - 1}"
-                });
-            }
-
-            if (!varParamComp.CanRemoveParameter(side, removeIndex))
-            {
-                return JsonConvert.SerializeObject(new
-                {
-                    success = false,
-                    error = $"Cannot remove parameter at index {removeIndex}",
-                    currentCount = paramList.Count,
-                    suggestion = "Component may require a minimum number of parameters"
-                });
-            }
-
-            var param = paramList[removeIndex];
-            string paramName = param.Name;
-
-            if (side == GH_ParameterSide.Input)
-                ghComponent.Params.UnregisterInputParameter(param);
-            else
-                ghComponent.Params.UnregisterOutputParameter(param);
-
-            varParamComp.VariableParameterMaintenance();
-            ghComponent.ExpireSolution(true);
-            doc.NewSolution(false);
-
-            return JsonConvert.SerializeObject(new
-            {
-                success = true,
-                id = ghComponent.InstanceGuid.ToString(),
-                operation = "remove",
-                side = side.ToString().ToLower(),
-                index = removeIndex,
-                removedParameterName = paramName,
-                newCount = paramList.Count
-            });
-        }
-
-        private string SetParameterCount(IGH_Component ghComponent, IGH_VariableParameterComponent varParamComp,
-            GH_ParameterSide side, int targetCount, GH_Document doc)
-        {
-            var paramList = side == GH_ParameterSide.Input ? ghComponent.Params.Input : ghComponent.Params.Output;
-            int currentCount = paramList.Count;
-
-            if (targetCount < 0)
-            {
-                return ToolHelpers.ErrorResponse("Target count must be non-negative");
-            }
-
-            if (targetCount == currentCount)
-            {
-                return JsonConvert.SerializeObject(new
-                {
-                    success = true,
-                    id = ghComponent.InstanceGuid.ToString(),
-                    operation = "set_count",
-                    side = side.ToString().ToLower(),
-                    count = currentCount,
-                    message = "Already at target count"
-                });
-            }
-
-            int added = 0, removed = 0;
-            var errors = new List<string>();
-
-            try
-            {
-                // Remove excess parameters
-                while (paramList.Count > targetCount)
-                {
-                    int removeIdx = paramList.Count - 1;
-                    if (!varParamComp.CanRemoveParameter(side, removeIdx))
-                    {
-                        errors.Add($"Cannot remove parameter at index {removeIdx}");
-                        break;
-                    }
-
-                    var param = paramList[removeIdx];
-                    if (side == GH_ParameterSide.Input)
-                        ghComponent.Params.UnregisterInputParameter(param);
-                    else
-                        ghComponent.Params.UnregisterOutputParameter(param);
-                    removed++;
-                }
-
-                // Add missing parameters
-                while (paramList.Count < targetCount)
-                {
-                    int insertIdx = paramList.Count;
-                    if (!varParamComp.CanInsertParameter(side, insertIdx))
-                    {
-                        errors.Add($"Cannot add parameter at index {insertIdx}");
-                        break;
-                    }
-
-                    var newParam = varParamComp.CreateParameter(side, insertIdx);
-                    if (newParam == null)
-                    {
-                        errors.Add($"Failed to create parameter at index {insertIdx}");
-                        break;
-                    }
-
-                    if (side == GH_ParameterSide.Input)
-                        ghComponent.Params.RegisterInputParam(newParam);
-                    else
-                        ghComponent.Params.RegisterOutputParam(newParam);
-                    added++;
-                }
-
-                varParamComp.VariableParameterMaintenance();
-                ghComponent.ExpireSolution(true);
-                doc.NewSolution(false);
-
-                bool success = paramList.Count == targetCount;
-                return JsonConvert.SerializeObject(new
-                {
-                    success,
-                    id = ghComponent.InstanceGuid.ToString(),
-                    operation = "set_count",
-                    side = side.ToString().ToLower(),
-                    targetCount,
-                    actualCount = paramList.Count,
-                    added,
-                    removed,
-                    errors = errors.Count > 0 ? errors : null
-                });
-            }
-            catch (Exception ex)
-            {
-                return ToolHelpers.ErrorResponse($"Failed during set_count: {ex.Message}");
-            }
-        }
-
-        #endregion
+        // Value Adjustment Actions - see GhCanvasTool.Values.cs
+        // Group Actions - see GhCanvasTool.Groups.cs
+        // Zoomable Actions - see GhCanvasTool.Zoomable.cs
     }
 }
